Beginning real implementation Kalman

+++ Initial_Material/controllers/localization_controller/kalman.h
@@ -1,7 +1,7 @@
 #ifndef KALMAN_H
 #define KALMAN_H 
 
-void init_kalman(float initial_x, float initial_y);
-void update_pos_kalman(position_t* pos, float acc[2], float meas_val[2], bool meas_true);
+void init_kalman(double initial_x, double initial_y);
+void update_pos_kalman(position_t* pos, double acc[2], double meas_val[2], bool meas_true);
 
 #endif
